Move MessageSink interface.
