+++ jetty-websocket/websocket-util/src/main/java/org/eclipse/jetty/websocket/util/messages/MessageSink.java
@@ -16,7 +16,7 @@
 // ========================================================================
 //
 
-package org.eclipse.jetty.websocket.util;
+package org.eclipse.jetty.websocket.util.messages;
 
 import org.eclipse.jetty.util.Callback;
 import org.eclipse.jetty.websocket.core.Frame;
